23.03.2016 17:50 In th middle of re-working rotational movement: keyboard seems to be glitching and reversed. IAngle support for Actor that enables feedback from Rotation strategy. Implementation of clickability. New game message broadcast options, debug statically connected to Log and info for player.

+++ Space Impact/Core/Game/AbstractPlayer.cs
@@ -43,8 +43,8 @@ namespace Space_Impact.Core.Game.Player
 			{
 				Position position = new Position();
 				//Moves to the player's center coordinates
-				position.X = X + (int)Width / 2;
-				position.Y = Y + (int)Height / 2;
+				position.X = X + (float)Width / 2;
+				position.Y = Y + (float)Height / 2;
 				return position;
 			}
 		}
