+++ Space Impact/Core/Graphics/IPlacedOnField.cs
@@ -8,8 +8,8 @@ namespace Space_Impact.Core
 {
 	public interface IPlacedOnField
 	{
-		int X { get; set; }
-		int Y { get; set; }
+		float X { get; set; }
+		float Y { get; set; }
 		double Width { get; }
 		double Height { get; }
 	}
